Update
deleting

+++ todo.php
@@ -22,28 +22,34 @@
 
 	foreach ($new_array as $value) {
 		print_r($value["listname"]);
-		echo "<button class='deletebutton' value='$list_id'>x</button>";
-		echo "<br>";
 
+		//deletes third list, how do we store the right list id
 		$list_id = $value["list_id"];
 
+		echo "<form method='post' action'todo.php'>";
+		echo "<input type='submit' name='deletelist' value='x'/>";
+		echo "<input type='hidden' name='id' value='$list_id'/>";
+		echo "</form>";
+
 		//display all tasks
-		$sql2 = "SELECT taskname FROM tasks WHERE list_id = '$list_id' "; //and user_id = 'username'";
-		$result2 = mysqli_query($db, $sql2);
+		//$sql2 = "SELECT taskname FROM tasks WHERE list_id = '$list_id' "; //and user_id = 'username'";
+		//$result2 = mysqli_query($db, $sql2);
 
 		$taskname = "";
 		$task_id = "";
 
-		/*/$stmt = $db->prepare("SELECT task_id FROM tasks WHERE list_id = '$list_id' ");
+		$stmt = $db->prepare("SELECT task_id, taskname FROM tasks WHERE list_id = '$list_id' ");
 		$stmt->execute();
-		$stmt->bind_result($task_id);/*/
+		$stmt->bind_result($task_id, $taskname);
 
-		if (mysqli_num_rows($result2) > 0) {
+		if (mysqli_num_rows($taskname) > 0) {
 
 				// output data of each row
-				while($row2 = mysqli_fetch_assoc($result2)) {
-						echo "". $row2["taskname"]. "<button class='deletebutton' value='$task_id'>x</button>";
-						echo "<br><br>";
+				while($row2 = mysqli_fetch_assoc($taskname)) {
+						echo "". $row2["taskname"]. "";
+
+						echo "<br>";
+
 				}
 			}
 		}
@@ -101,32 +107,50 @@
 	}
 
 	//Remove list and tasks with same list_id if deletbutton is clicked
-	/*/
-				$stmt = $db->prepare ("DELETE FROM lists WHERE list_id = '$list_id'");
-        $stmt->bind_param('i', $list_id);
-        $response = $stmt->execute();
-        printf("<br>List deleted!");
-	/*/
-  //Remove finished tasks
+
+	if (isset($_POST['deletelist'])) {
+
+		$stmt = $db->prepare ("DELETE FROM lists WHERE list_id = '$list_id'");
+		//$stmt->bind_param('i', $list_id);
+		//$response = $stmt->execute();
+		$stmt->execute();
+
+		$stmt = $db->prepare ("DELETE FROM tasks WHERE list_id = '$list_id'");
+		//$stmt->bind_param('i', $list_id);
+		//$response = $stmt->execute();
+		$stmt->execute();
+
+		header("Refresh:0");
+  }
+
+
+//Remove specific task in a list if deletebutton is clicked
 
 ?>
 
+<h2>Add new list</h2>
 <form action="todo.php" method="POST">
+	<h3>List name</h3>
 	<input type="text" name="newlist" placeholder="New List" class="inputField">
 	<br>
-	<input type="submit" name="submitlist" value="Add" class="button">
+	<input type="submit" name="submitlist" value="Create" class="button">
 </form>
 
+<h2>Add new task</h2>
 <form action="todo.php" method="POST">
+	<h3>Task name</h3>
 	<input type="text" name="newtask" placeholder="Add Task" class="inputField">
 	<br>
+	<h3>Task description</h3>
 	<input type="text" name="newtaskdesc" placeholder="Task Description" class="inputField">
 	<br>
+	<h3>Start date</h3>
 	<input type="date" name="newStartDate" class="inputField">
 	<br>
+	<h3>End date</h3>
 	<input type="date" name="newEndDate" class="inputField">
 	<br>
-
+	<h3>Select list</h3>
   <?php
 
   $sql3 = "SELECT list_id,listname FROM lists";
